# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API Documentation",
        "description": "Swagger documentation for my Laravel project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/podcasts": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Get all podcasts",
                "operationId": "28abe6e0a78fe6785659a51dde99679e",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Create a new podcast",
                "operationId": "e8917ecbdaa178c760e0992b16b91fb1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "My Podcast"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Podcast description"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Podcast created successfully"
                    }
                }
            }
        },
        "/api/podcasts/{id}": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Get a specific podcast",
                "operationId": "9009d1f97f328b3e48f85075f935aa0f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Podcast ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Podcast not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Update a podcast",
                "operationId": "e72afa88b52a2b4864a2e29cca8cdabc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Podcast ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Podcast updated successfully"
                    },
                    "404": {
                        "description": "Podcast not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Delete a podcast",
                "operationId": "378754b8c72fa0738edc3691e12cfb13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Podcast ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podcast deleted successfully"
                    },
                    "404": {
                        "description": "Podcast not found"
                    }
                }
            }
        },
...
    },
...
}